Fix delay
Changed the delay argument to be the time between setting inputs and getting outputs. Also, fixed the mod not actually waiting for the delay period.

+++ src/main/java/cmsc389e/circuitry/common/command/CommandCircuitryBase.java
@@ -38,7 +38,7 @@ public abstract class CommandCircuitryBase extends CommandBase {
 	    if (type.equals(Integer.class))
 		return parseInt(arg);
 	    if (type.equals(Long.class))
-		return parseDouble(arg);
+		return parseLong(arg);
 	}
 	if (type.equals(String.class))
 	    return arg;
